+++ src/App.js
@@ -1,8 +1,8 @@
 import { BrowserRouter, Route, Routes } from "react-router-dom";
 
 import RandingPage from "./component/main/RandingPage";
-import ScriptBoard from "./component/ScriptBoard";
-import ScriptBoardSelect from "./component/ScriptBoardSelect";
+import ScriptBoard from "./component/ScripBoard/ScriptBoard";
+import ScriptBoardSelect from "./component/ScripBoard/ScriptBoardSelect";
 
 const App = () => {
   return (
