+++ src/component/ScripBoard/WriteButton.jsx
