+++ src/component/TextPreview/TextPreview.css
