+++ src/component/TextPreview/TextPreviewList.css
